Criei o componente Dashboard como template
